requirements.txt
boto3==1.29.2
botocore==1.32.2
branca==0.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
folium==0.15.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
numpy==1.26.2
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
urllib3==2.0.7
google-cloud-storage